requirements.txt
Flask
Flask-SQLAlchemy
pandas
openpyxl
python-docx
reportlab
gunicorn
psycopg2-binary
Flask-Migrate
python-dotenv